refactor(api): 抽取启动索引初始化到独立模块，main.py 回到行数阈值内
